RPA
This is an RPA bot for Reading Domestic and international travel requests the bot will also read the manager's reply for Request Approval or Rejected update the whole data in the database and generate monthly reports for the entire process. candidates, International and Domestic charts and traveller charts created in sheets.